reggie: fixes for continued building

+++ regressioncheck/regressioncheck_run.f90
@@ -442,8 +442,8 @@ END SUBROUTINE CheckExampleName
 
 
 !===================================================================================================================================
-!> IF build-mode: read the configurations.reggie and determine the number of builds
-!> ELSE         : nReggieBuilds=1
+!> IF BuildSolver=T: read the configurations.reggie and determine the number of builds
+!> ELSE            : nReggieBuilds=1
 !===================================================================================================================================
 SUBROUTINE GetnReggieBuilds(iExample,ReggieBuildExe,N_compile_flags,nReggieBuilds)
 !===================================================================================================================================
@@ -451,8 +451,7 @@ SUBROUTINE GetnReggieBuilds(iExample,ReggieBuildExe,N_compile_flags,nReggieBuild
 ! MODULES
 USE MOD_Globals
 USE MOD_RegressionCheck_Build,   ONLY: ReadConfiguration
-USE MOD_RegressionCheck_Vars,    ONLY: BuildDir
-USE MOD_RegressionCheck_Vars,    ONLY: BuildCounter,BuildSolver
+USE MOD_RegressionCheck_Vars,    ONLY: BuildDir,BuildCounter,BuildSolver,BuildContinue
 ! IMPLICIT VARIABLE HANDLING
 IMPLICIT NONE
 !-----------------------------------------------------------------------------------------------------------------------------------
@@ -473,10 +472,12 @@ IF(BuildSolver)THEN
   IF(ReggieBuildExe.EQ.'')THEN
     CALL ReadConfiguration(iExample,nReggieBuilds,N_compile_flags)
     BuildCounter=1 ! reset the counter between read and build (used for selecting the build configuration for compilation)
-    SYSCOMMAND='rm -rf '//TRIM(BuildDir)//'build_reggie_bin > /dev/null 2>&1'
-    CALL EXECUTE_COMMAND_LINE(SYSCOMMAND, WAIT=.TRUE., EXITSTAT=iSTATUS)
-    SYSCOMMAND= 'mkdir '//TRIM(BuildDir)//'build_reggie_bin'
-    CALL EXECUTE_COMMAND_LINE(SYSCOMMAND, WAIT=.TRUE., EXITSTAT=iSTATUS)
+    IF(.NOT.BuildContinue)THEN ! only delete the executables folder if reggie starts at the beginning
+      SYSCOMMAND='rm -rf '//TRIM(BuildDir)//'build_reggie_bin > /dev/null 2>&1'
+      CALL EXECUTE_COMMAND_LINE(SYSCOMMAND, WAIT=.TRUE., EXITSTAT=iSTATUS)
+      SYSCOMMAND= 'mkdir '//TRIM(BuildDir)//'build_reggie_bin'
+      CALL EXECUTE_COMMAND_LINE(SYSCOMMAND, WAIT=.TRUE., EXITSTAT=iSTATUS)
+    END IF
   ELSE
     SWRITE(UNIT_stdOut,'(A)')'regressioncheck_run.f90: CALL ReadConfiguration() has already been performed, skipping...'
   END IF
@@ -501,6 +502,7 @@ USE MOD_RegressionCheck_Vars,    ONLY: BuildDir,CodeNameLowCase,EXECPATH
 USE MOD_RegressionCheck_Vars,    ONLY: BuildValid,BuildSolver
 USE MOD_RegressionCheck_Tools,   ONLY: CheckForExecutable
 USE MOD_RegressionCheck_Build,   ONLY: BuildConfiguration
+USE MOD_RegressionCheck_Vars,    ONLY: BuildContinue,BuildContinueNumber
 ! IMPLICIT VARIABLE HANDLING
 IMPLICIT NONE
 !-----------------------------------------------------------------------------------------------------------------------------------
@@ -521,15 +523,28 @@ SkipBuild=.FALSE.
 ExitBuild=.FALSE.
 ! Get code binary (build or find it)
 IF(BuildSolver)THEN
+!print*,""
+!print*,"iReggieBuild=[",iReggieBuild,"]"
+!print*,"BuildValid(iReggieBuild)=[",BuildValid(iReggieBuild),"]"
+
   ! if build is not valid no binary has been built and the lopp can cycle here
   IF(.NOT.BuildValid(iReggieBuild))THEN ! invalid reggie build 
-    WRITE (ReggieBuildExe, '(a, i4.4)') "invalid"
+    WRITE (ReggieBuildExe, '(a, i4.4)') "[BUILD_is_invalid]"
   ELSE
     WRITE (ReggieBuildExe, '(a, i4.4)') CodeNameLowCase, COUNT(BuildValid(1:iReggieBuild)) ! e.g. XXXXX0001
   END IF
   ! check if build exists -> if it does, don't build a new executable with cmake
   FileName=TRIM(BuildDir)//'build_reggie_bin/'//ReggieBuildExe
+!print*,"FileName=["//TRIM(FileName)//"]"
   INQUIRE(File=FileName,EXIST=ExistFile)
+
+  IF(BuildContinue)THEN ! even if the executable exists, re-build it when it equels BuildContinueNumber
+    IF(COUNT(BuildValid(1:iReggieBuild)).EQ.BuildContinueNumber)THEN
+      ExistFile=.FALSE.
+    END IF
+  END IF
+
+
   IF(ExistFile) THEN ! 1. build already exists (e.g. XXXX0001 located in ../build_reggie_bin/)
     EXECPATH=TRIM(FileName)
   ELSE ! 2. build does not exists -> create it
@@ -540,6 +555,9 @@ IF(BuildSolver)THEN
       EXECPATH=TRIM(BuildDir)//'build_reggie_bin/'//ReggieBuildExe
     END IF
   END IF
+
+
+
   ! if build is not valid no exe has been built and the lopp can cycle here
   IF((.NOT.BuildValid(iReggieBuild)).AND.(iReggieBuild.NE.nReggieBuilds))THEN ! invalid reggie build but not last reggie build
     SkipBuild=.TRUE. ! -> does a CYLCE
@@ -547,6 +565,19 @@ IF(BuildSolver)THEN
                                                                                   ! ("cycle" would start an infinite loop)
     ExitBuild=.TRUE. ! -> does an EXIT
   END IF
+
+
+  IF(BuildContinue)THEN ! skip build if the build is smaller than BuildContinueNumber
+    IF(COUNT(BuildValid(1:iReggieBuild)).LT.BuildContinueNumber)THEN ! skip the first N builds when BuildContinue is used
+    !IF(iReggieBuild.LT.BuildContinueNumber)THEN ! skip the first N builds when BuildContinue is used
+!print*,"COUNT(BuildValid(1:iReggieBuild))",COUNT(BuildValid(1:iReggieBuild))
+!print*,"BuildContinueNumber              ",BuildContinueNumber,"----> RETURN"
+!read*
+      SkipBuild=.TRUE.
+      !RETURN
+    END IF
+  END IF
+
   CALL CheckForExecutable(Mode=2) ! check if executable was created correctly
 END IF
 END SUBROUTINE GetCodeBinary
